Allow range operator
If the query string contains range boundaries, i.e.  price>15,price<20, the latter predicate will overwrite the former one.
So here inside the check for the field, there is another check if we previously added a predicate to the same operator. If so, we just add to it, rather than overwriting it.